open file

+++ src/Components/DocumentTable.jsx
@@ -1,14 +1,18 @@
 import * as React from 'react';
-import { View, ActivityIndicator } from 'react-native'
+import { View, ActivityIndicator, Platform, Alert } from 'react-native'
 
 import { DataTable, Text } from 'react-native-paper';
 import { format } from 'date-fns'
 import { Feather } from '@expo/vector-icons';
 import Toast from 'react-native-root-toast';
+import * as FileSystem from 'expo-file-system';
+import * as IntentLauncher from 'expo-intent-launcher';
 
 import instance from '../api/axios';
 
-const Document = ({ items }) => {
+const Document = ({ items: initialItems }) => {
+  const [items, setItems] = React.useState(initialItems);
+
   const [page, setPage] = React.useState(0);
   const [numberOfItemsPerPageList] = React.useState([2, 5, 10]);
   const [itemsPerPage, onItemsPerPageChange] = React.useState(
@@ -24,27 +28,62 @@ const Document = ({ items }) => {
   }, [itemsPerPage]);
 
   const deleteDocument = React.useCallback(async (id) => {
-    setIsLoading(true);
 
+    Alert.alert('Delete Action', 'Are you sure you want to delete?', [
+      {
+        text: 'Cancel',
+        onPress: () => console.log('Cancel Pressed'),
+        style: 'cancel',
+      },
+      {
+        text: 'OK',
+        onPress: async () => {
+          setIsLoading(true);
+
+          try {
+            const res = await instance.delete(
+              `/api/booking/delete-document/${id}`,
+            );
+
+            const { type } = res.data;
+            setIsLoading(false);
+
+            if (type === 'success') {
+              setItems((prev) => prev.filter((doc) => doc.id !== id))
+              Toast.show('Document deleted successfully');
+            }
+          } catch (error) {
+            setIsLoading(false);
+            console.error(error.response?.data);
+
+            const { message } = error.response?.data;
+
+            Toast.show(message ?? 'Something went wrong.', {
+              duration: Toast.durations.LONG,
+              textColor: 'red'
+            });
+          }
+        }
+      },
+    ]);
+
+  }, [items]);
+
+  const handleFilePress = React.useCallback(async (doc) => {
     try {
-      const res = await instance.delete(
-        `/api/booking/delete-document/${id}`,
-      );
+      const { uri, mimeType } = await FileSystem.downloadAsync(doc.full_document_path)
 
-      const { success } = res.data;
-      setIsLoading(false);
-
-      if (success) {
-        Toast.show('Document deleted successfully');
+      if (Platform.OS !== 'ios') {
+        IntentLauncher.startActivityAsync('android.intent.action.VIEW', {
+          data: uri,
+          flags: 1,
+          type: mimeType
+        });
       }
-    } catch (error) {
-      setIsLoading(false);
-      console.error(error.response?.data);
+    } catch (e) {
+      console.error(e);
 
-      const { message } = error.response?.data;
-
-      Toast.show(message ?? 'Something went wrong.', {
-        duration: Toast.durations.LONG,
+      Toast.show('Something went wrong!', {
         textColor: 'red'
       });
     }
@@ -68,7 +107,7 @@ const Document = ({ items }) => {
 
       {items.length > 0 ? (
         items.slice(from, to).map((item) => (
-          <DataTable.Row key={item.id}>
+          <DataTable.Row key={item.id} onPress={async () => await handleFilePress(item)}>
             <DataTable.Cell>{item.name}</DataTable.Cell>
             <DataTable.Cell numeric>{format(item.date_added, 'dd-MM-yyyy')}</DataTable.Cell>
             <DataTable.Cell numeric>
